Releasing major version v3.0.0, It contains breaking changes please check changelog for the same.

+++ filter/number_filter.go
@@ -1,12 +1,12 @@
 package filter
 
 type NumberFilter struct {
-	Is        int32   `json:"is,omitempty"`
-	IsNot     int32   `json:"is_not,omitempty"`
-	Lt        int32   `json:"lt,omitempty"`
-	Lte       int32   `json:"lte,omitempty"`
-	Gt        int32   `json:"gt,omitempty"`
-	Gte       int32   `json:"gte,omitempty"`
-	Between   []int32 `json:"between,omitempty"`
+	Is        int64   `json:"is,omitempty"`
+	IsNot     int64   `json:"is_not,omitempty"`
+	Lt        int64   `json:"lt,omitempty"`
+	Lte       int64   `json:"lte,omitempty"`
+	Gt        int64   `json:"gt,omitempty"`
+	Gte       int64   `json:"gte,omitempty"`
+	Between   []int64 `json:"between,omitempty"`
 	IsPresent *bool   `json:"is_present,omitempty"`
 }
